krypto.md, KRYPTO.ml: Move Hash functions to own module
Since the hash functions depend only on Cryptokit, this reduces the dependencies
needed when only using that functionality.

We also add some CSS selectors to the markdown file to allow conditionally
including just the Hash functionality.

Aliases are added for the old hooks, but they are only intended for backward
compatibility.